The role of demographics, underlying clinical conditions, and governmental policies during the COVID-19 Pandemic.
In this project, focus was on mask mandates and vaccination numbers to visually connect the dramatic changes to COVID19 national morbidity and mortality rate. This project's attention to datasets helped to make sense of the numbers by analyzing the most and least affected ethnic groups. Also reviewed is how government facemark mandates influenced prevalence and death toll. The virus seemed to capture life, and the health care community couldn't agree on how to curtail or prevent COVID-19's wide variety of health conditions.

This project found out that facemark mandate did not curtail the spread of COVID19